requirements.txt
flask
gunicorn
numpy
opencv-python-headless
tensorflow-cpu==2.12.0
flask-cors
pillow